LPS-85787 Merge SortParser into SortProvider. SortProvider is the only user of SortParser and the impl is hardcoded, it should not be a separate service since it's only deligating. SortParser doesn't make sense as a name because it's really a SortFieldParser, deprecate the interface. Rename SortParserRegistry to SortProviderRegistrar because it doesn't hold the reference, only control it - we're using OSGi as the Registry.
