+++ modules/apps/headless-apio/structured-content/structured-content-apio-impl/src/main/java/com/liferay/structured/content/apio/internal/architect/provider/SortProvider.java
@@ -15,21 +15,31 @@
 package com.liferay.structured.content.apio.internal.architect.provider;
 
 import com.liferay.apio.architect.provider.Provider;
+import com.liferay.petra.string.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
+import com.liferay.structured.content.apio.architect.entity.EntityField;
+import com.liferay.structured.content.apio.architect.entity.EntityModel;
+import com.liferay.structured.content.apio.architect.sort.InvalidSortException;
 import com.liferay.structured.content.apio.architect.sort.Sort;
-import com.liferay.structured.content.apio.architect.sort.SortParser;
-import com.liferay.structured.content.apio.internal.architect.filter.StructuredContentEntityModel;
+import com.liferay.structured.content.apio.architect.sort.SortField;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import javax.servlet.http.HttpServletRequest;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Cristina González
  */
-@Component(service = Provider.class)
 public class SortProvider implements Provider<Sort> {
+
+	public SortProvider(EntityModel entityModel) {
+		_entityModel = entityModel;
+	}
 
 	@Override
 	public Sort createContext(HttpServletRequest httpServletRequest) {
@@ -39,12 +49,110 @@ public class SortProvider implements Provider<Sort> {
 			return Sort.emptySort();
 		}
 
-		return new Sort(_sortParser.parse(sortString));
+		return new Sort(parse(sortString));
 	}
 
-	@Reference(
-		target = "(entity.model.name=" + StructuredContentEntityModel.NAME + ")"
-	)
-	private SortParser _sortParser;
+	/**
+	 * Returns a List of {@link SortField} obtained from a comma-separated list
+	 * of field names and sort directions.
+	 *
+	 * Sort directions supported are desc and asc and can be appended to each
+	 * sort field, separated by the ':' character.
+	 *
+	 * If a sort direction is not provided, the sort field will be 'asc'.
+	 *
+	 * For example:
+	 * - field1,field2,field3
+	 * - field1:asc,field2:desc,field3
+	 * - field1:asc,field2,field3:desc
+	 *
+	 * @param  sortString - String to be parsed
+	 * @return a  {@link List <SortField>}
+	 * @review
+	 */
+	public List<SortField> parse(String sortString) {
+		if (Validator.isNull(sortString)) {
+			return Collections.emptyList();
+		}
+
+		List<String> list = StringUtil.split(sortString);
+
+		Stream<String> stream = list.stream();
+
+		return stream.map(
+			this::getSortFieldOptional
+		).flatMap(
+			sortFieldOptional -> sortFieldOptional.map(
+				Stream::of
+			).orElseGet(
+				Stream::empty
+			)
+		).collect(
+			Collectors.toList()
+		);
+	}
+
+	protected Optional<SortField> getSortFieldOptional(String sortString) {
+		List<String> list = StringUtil.split(sortString, ':');
+
+		if (list.isEmpty()) {
+			return Optional.empty();
+		}
+
+		if (list.size() > 2) {
+			throw new InvalidSortException(
+				"Unable to parse sort string: " + sortString);
+		}
+
+		String fieldName = list.get(0);
+
+		boolean ascending = _ASC_DEFAULT;
+
+		if (list.size() > 1) {
+			ascending = isAscending(list.get(1));
+		}
+
+		Map<String, EntityField> entityFieldsMap =
+			_entityModel.getEntityFieldsMap();
+
+		EntityField entityField = entityFieldsMap.get(fieldName);
+
+		if (entityField == null) {
+			throw new InvalidSortException(
+				"Unable to sort by field: " + fieldName);
+		}
+
+		return Optional.of(new SortField(entityField, ascending));
+	}
+
+	protected boolean isAscending(String orderBy) {
+		if (orderBy == null) {
+			return _ASC_DEFAULT;
+		}
+
+		if (_ORDER_BY_ASC.equals(
+				com.liferay.portal.kernel.util.StringUtil.toLowerCase(
+					orderBy))) {
+
+			return true;
+		}
+
+		if (_ORDER_BY_DESC.equals(
+				com.liferay.portal.kernel.util.StringUtil.toLowerCase(
+					orderBy))) {
+
+			return false;
+		}
+
+		return _ASC_DEFAULT;
+	}
+
+	private static final boolean _ASC_DEFAULT = true;
+
+	private static final String _ORDER_BY_ASC = "asc";
+
+	private static final String _ORDER_BY_DESC = "desc";
+
+	private final EntityModel _entityModel;
 
 }
